Implemented the ITEST mode. Current control can follow a +-200mA 100 Hz square wave fairly well. There exists large negative iMeasured spikes which negatively affect tracking- this will need to be addressed in another commit.

+++ iSense.h
@@ -3,6 +3,6 @@
 
 void initADC();
 int getADCcount();
-int getCurrentmAmps();
+float getCurrentmAmps();
 
 #endif
